Fix DateOffset week counting and int division in mean_in_*

+++ ivory-operation/src/main/scala/com/ambiata/ivory/operation/extraction/reduction/MeanInDaysReducer.scala
@@ -9,7 +9,7 @@ object MeanInReducer {
       sum += i
       count += 1
     }
-    if (count == 0) 0 else sum / count
+    if (count == 0) 0 else sum.toDouble / count
   }
 }
